Moving validation infrastructure to qiskit.validation
